python_app/requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
requests==2.31.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1 